Create README.md
The app helps users search for plants in stock at a plant nursery and also provides a shopping cart so users can determine the overall cost of their purchases. The app will be fully responsive for all screen sizes. I envision it on customer computer consoles in the nursery. I also envision users using the app on their smartphones as they are browsing the store.